requirements.txt
Django==3.2.6

asgiref==3.4.1
pip==22.1.1
pytz==2021.3
setuptools==47.1.0
sqlparse==0.4.2
typing-extensions==4.0.1
atomicwrites==1.4.0
attrs==21.4.0
beautifulsoup4==4.10.0
colorama==0.4.4
fontawesomefree==5.15.4
importlib-metadata==4.10.1
iniconfig==1.1.1
mod-wsgi==4.9.0
packaging==21.3
Pillow==9.0.0
pluggy==1.0.0
py==1.11.0
pyodbc==4.0.32
pyparsing==3.0.7
pypi==2.1
pytest==6.2.5
soupsieve==2.3.1
toml==0.10.2
wheel==0.37.1
zipp==3.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
django-filter==21.1
django-multiselectfield==0.1.12
idna==3.3
oauthlib==3.1.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
urllib3==1.26.8
django-environ==0.8.1
psycopg2==2.9.3
install==1.3.5
tmdbv3api==1.7.6
typing_extensions==4.0.1
django-jsonify==0.3.0
astroid==2.9.3
django-embed-video==1.4.1
django-jsonfield==1.4.1
djangorestframework==3.13.1
isort==5.10.1
jsonfield==3.1.0
lazy-object-proxy==1.7.1
mccabe==0.6.1
platformdirs==2.5.1
pylint==2.12.2
pylint-django==2.5.2
pylint-plugin-utils==0.7
six==1.16.0
typed-ast==1.5.2
wrapt==1.13.3
django-bootstrap-modal-forms==2.2.0
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
django-celery==3.3.1
kombu==3.0.37
django-cors-headers==3.13.0
exceptiongroup==1.0.0
tomli==2.0.1
